+++ .gitignore
@@ -1,5 +1,6 @@
 # Python temporary files
 __pycache__
+Pipfile.lock
 
 # Logs
 log
